qemu in libafl_sugar
