+++ libafl_sugar/Cargo.toml
@@ -9,4 +9,5 @@ edition = "2018"
 [dependencies]
 libafl = { path = "../libafl", version = "0.5.0" }
 libafl_targets = { path = "../libafl_targets", version = "0.5.0" }
+libafl_qemu = { path = "../libafl_qemu", version = "0.5.0" }
 typed-builder = "0.9.0" # Implement the builder pattern at compiletime
